fs: namespace: revert sus_mount to v1.5.5
SUSFS v1.5.6-v1.5.7 introduced mount ID reordering optimizations. But since it's for GKI or a specific fix for some devices, in this device it could
cause more harm than good which it leaks mount-related detections.

For now revert the sus_mount mechanism to v1.5.5 as it's still okay and stable.